Mainly allow KDF iters/rounds to be specified with -R, but ...:
Also change the default iterations for SHA-* KDF to 100 000 and the
default rounds for bcrypt to 36. This breaks database opening with older
databases, hence (mostly) the new '-R' option.

Sync cmd_import.c/cmd_export.c/cmd_passwd.c with kc.c's option parsing,
parameter parsing and related stuff.

Switch the kc_ssha_get_password() and kc_ykchalresp() calls so as to
avoid a too long password to be used with kc_ykchalresp() after
kc_ssha_get_password(). Also error out when using both but only one of
them has the ",password" parameter specified.

Fix 'status' command's "Password: yes/no" output by taking both -y and
-A into account and checking their ",password" parameters.

Add some regression tests for the '-R' option.

Extend manual with the new options and rewrite some old ones.

+++ regress/cmd_export.sh
@@ -41,6 +41,22 @@ else
 fi
 
 
+rm -f regress/test_export.kcd
+printf "export -R 16 -P bcrypt -k regress/test_export\n${PASSWORD}\n${PASSWORD}\n" |${KC_RUN} -b -k ${KC_DB} -p ${KC_PASSFILE}
+
+if [ ! -r "regress/test_export.kcd" ];then
+	echo "$0 test failed (unreadable export file, bcrypt (16 rounds))!"
+	exit 1
+fi
+
+if printf "${PASSWORD}" |${KC_RUN} -b -k regress/test_export.kcd -P bcrypt -R 16;then
+	echo "$0 test ok (export, bcrypt (16 rounds))!"
+else
+	echo "$0 test failed (export, bcrypt (16 rounds))!"
+	exit 1
+fi
+
+
 if [ ${SCRYPT} ];then
 	rm -f regress/test_export.kcd
 	printf "export -k regress/test_export -P scrypt\n${PASSWORD}\n${PASSWORD}\n" |${KC_RUN} -b -k ${KC_DB} -p ${KC_PASSFILE}
